Revised the Windows build system in the following manner:
   1. Introduced a peace_library project to build a static library just like
      it is done under Windows.
   2. Revied the peace project to utilize the library
   3. Introduced a new project for building examples under Windows as per
      Karro's request.

Fixed a couple of minor type mismatch warnings that vc9 was unhappy about
(size_t to int conversions)

Added a check to handle a sentinel EST file name called "<none>" that can
be set by programs to specify a dummy EST file to simply suppress error that
a EST file has not been specified. This is used from UsingPeace.cpp program.

Added a simple example code to UsingPeace.cpp to illustrate how parameters
can be set for d2.

+++ src/BatonESTData.cpp
@@ -30,7 +30,7 @@ BatonESTData::BatonESTData(const char* estSequence, const int sectionWidth,
                            const int nMer, const int maxMer) {
     maxMerValue = maxMer;
     section = sectionWidth;
-    size = strlen(estSequence) - (nMer - 1);
+    size = (int) strlen(estSequence) - (nMer - 1);
     nSections = size / section;
     int hangover = size%section;
     if (hangover > section/2)
